+++ package.json
@@ -17,7 +17,7 @@
 		"test": "vitest run",
 		"test:watch": "vitest",
 		"cap:sync": "cap sync",
-		"build:native": "BUILD_TARGET=native vite build"
+		"build:native": "BUILD_TARGET=native vite build && cap sync"
 	},
 	"devDependencies": {
 		"@capacitor/cli": "^7",
