+++ src/app.css
@@ -8,6 +8,15 @@
 
 @custom-variant dark (&:is(.dark *));
 
+@theme {
+	/* Safe-area offsets for native (notch/status/nav bars), with the usual 0.5rem
+	   gutter folded in. Lives in the spacing namespace, so it works with any
+	   inset/padding utility: top-safe-inset-top, bottom-safe-inset-bottom,
+	   pb-safe-inset-bottom, etc. */
+	--spacing-safe-inset-top: calc(0.5rem + env(safe-area-inset-top));
+	--spacing-safe-inset-bottom: calc(0.5rem + env(safe-area-inset-bottom));
+}
+
 :root {
 	--background: oklch(1 0 0);
 	--foreground: oklch(0.145 0 0);
