+++ src/components/ui/Toast.svelte
@@ -5,7 +5,7 @@
 	import { X } from "lucide-svelte";
 </script>
 
-<div class="top-[calc(0.5rem+env(safe-area-inset-top))] flex w-full justify-center px-2 pointer-events-none fixed z-100">
+<div class="top-safe-inset-top flex w-full justify-center px-2 pointer-events-none fixed z-100">
 	<div
 		transition:slide={{ duration: 70 }}
 		class="bg-background text-foreground border-border pointer-events-auto flex items-center justify-between gap-4 overflow-hidden rounded-md border pl-6 pr-2 py-2 shadow-lg transition-all"
