+++ src/components/ui/modal/ModalTop.svelte
@@ -21,7 +21,7 @@
 <Modal
 	{modalType}
 	{onopenchange}
-	class="data-[state=closed]:slide-out-to-top-10 data-[state=open]:slide-in-from-top-10! zoom-in-100! zoom-out-100! top-[calc(0.5rem+env(safe-area-inset-top))]! translate-y-0! {class_}"
+	class="data-[state=closed]:slide-out-to-top-10 data-[state=open]:slide-in-from-top-10! zoom-in-100! zoom-out-100! top-safe-inset-top! translate-y-0! {class_}"
 >
 	{@render children?.()}
 </Modal>
