+++ src/components/ui/popups/BasePopup.svelte
@@ -16,7 +16,7 @@
 		copyToClipboard,
 		hasClipboardWrite
 	} from "@/lib/utils/device";
-	import { getPublicOrigin } from "@/lib/native/runtime";
+	import { getRootOrigin } from "@/lib/native/runtime";
 	import { getLocale } from "@/lib/paraglide/runtime";
 	import * as m from "@/lib/paraglide/messages";
 
@@ -51,7 +51,7 @@
 	} = $props();
 
 	function getShareUrl() {
-		return getPublicOrigin() + getCurrentPath() + "?lang=" + getLocale();
+		return getRootOrigin() + getCurrentPath() + "?lang=" + getLocale();
 	}
 </script>
