+++ src/lib/native/nativeUi.ts
@@ -11,7 +11,7 @@ import { isNative } from "@/lib/native/runtime";
 export function installNativeUi(): void {
 	if (!isNative()) return;
 
-	// Kill any service worker registered by an earlier build — it intercepts and
+	// Kill any service worker registered by an earlier build, it intercepts and
 	// breaks cross-origin map tiles, and is unnecessary inside the native shell.
 	if ("serviceWorker" in navigator) {
 		navigator.serviceWorker
