+++ src/lib/services/uicons.svelte.ts
@@ -46,10 +46,6 @@ export async function initIconSet(id: string, url: string, thisFetch: typeof fet
 }
 
 export async function initAllIconSets(thisFetch: typeof fetch = fetch) {
-	// On native, asset URLs must be absolute (the instance origin): unlike the
-	// map's fetch-based image loads, UI <img> tags and CSS url() bypass the
-	// native fetch wrapper, so a relative "/assets/..." would resolve to the
-	// local webview origin and 404. getInstanceUrl() is "" on web (unchanged).
 	const base = getInstanceUrl();
 	await Promise.all(
 		getConfig().uiconSets.map((s) => initIconSet(s.id, `${base}/assets/${s.id}/`, thisFetch))
