+++ src/lib/utils/mapUrl.ts
@@ -6,11 +6,7 @@ function getLinkGoogleMaps(coords: Coords) {
 	return `https://maps.google.com?q=${coords.lat},${coords.lon}`;
 }
 
-/**
- * Generic geo: URI — lets the OS route to the user's default maps/navigation app
- * instead of forcing Google/Apple. Used on native, where the provider choice is hidden.
- */
-function getLinkGeo(coords: Coords, name: string | undefined) {
+function getLinkNative(coords: Coords, name: string | undefined) {
 	const query = name ? `${coords.lat},${coords.lon}(${name})` : `${coords.lat},${coords.lon}`;
 	return `geo:${coords.lat},${coords.lon}?q=${encodeURIComponent(query)}`;
 }
@@ -25,7 +21,7 @@ function getLinkAppleMaps(coords: Coords, name: string | undefined = undefined) 
 
 export function getMapsUrl(coords: Coords, name: string | undefined = undefined) {
 	if (isNative()) {
-		return getLinkGeo(coords, name);
+		return getLinkNative(coords, name);
 	}
 	if (getUserSettings().externalMapProvider === ExternalMapProvider.APPLE) {
 		return getLinkAppleMaps(coords, name);
